+++ test/cases/_preprocessor/ejs-include-md/webpack.config.js
@@ -1,0 +1,56 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('@test/html-bundler-webpack-plugin');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  resolve: {
+    alias: {
+      '@images': path.join(__dirname, '../../../fixtures/images'),
+    },
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: [
+        {
+          import: 'src/index.ejs',
+          filename: 'demo.html',
+          data: {
+            title: 'Demo',
+            markdownFile: '/src/markdown/demo.md', // path relative to project directory
+          },
+        },
+        {
+          import: 'src/index.ejs',
+          filename: 'code.html',
+          data: {
+            title: 'Code',
+            markdownFile: '/src/markdown/code.md', // path relative to project directory
+          },
+        },
+      ],
+      preprocessor: 'ejs',
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: ['css-loader'],
+      },
+      {
+        test: /\.(png|svg|jpe?g|webp)$/i,
+        type: 'asset/resource',
+        generator: {
+          filename: 'img/[name].[hash:8][ext]',
+        },
+      },
+    ],
+  },
+};